Receipt App show modal as printview. Relocated AltinnModal to Molecules. Documented AltinnModal printview variant.
