+++ src/react-apps/applications/shared/src/components/molecules/AltinnModal.tsx
@@ -2,7 +2,7 @@ import { createMuiTheme, createStyles, IconButton, Modal, Typography } from '@ma
 import { withStyles } from '@material-ui/core/styles';
 import classNames from 'classnames';
 import * as React from 'react';
-import altinnTheme from '../theme/altinnStudioTheme';
+import altinnTheme from './../../theme/altinnStudioTheme';
 
 export interface IAltinnModalComponentProvidedProps {
   /** @ignore */
